added formatter for round up

+++ StarlingTestApp/Feed/Services/SavingsGoals/SavingsGoalsEndpoint.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 enum SavingsGoalsEndpoint {
-    case addMoney(accountId: String, savingsGoalsId: String, count: Int)
+    case addMoney(accountId: String, savingsGoalsId: String, count: Int64)
 }
 
 extension SavingsGoalsEndpoint: PutEndpoint {
